+++ app/src/main/java/com/example/rfidtab/service/model/kitorder/KitOrderCards.kt
@@ -3,8 +3,8 @@ package com.example.rfidtab.service.model.kitorder
 import com.google.gson.annotations.SerializedName
 
 class KitOrderCards(
-    @SerializedName("pipeSerialNumber") val pipeSerialNumber: Long?,
-    @SerializedName("couplingSerialNumber") val couplingSerialNumber: Long?,
-    @SerializedName("serialNoOfNipple") val serialNoOfNipple: Long?,
+    @SerializedName("pipeSerialNumber") val pipeSerialNumber: Int?,
+    @SerializedName("couplingSerialNumber") val couplingSerialNumber: Int?,
+    @SerializedName("serialNoOfNipple") val serialNoOfNipple: Int?,
     @SerializedName("rfidTagNo") val rfidTagNo: String?
 )
